+++ CHANGELOG.md
@@ -30,6 +30,17 @@
 - Now requires Craft 5.0+.
 - Now requires Craft Commerce 5.0+.
 
+## 3.0.4 - 2024-12-14
+
+### Added
+- Add “From Name” and “From Email” plugin settings.
+
+### Changed
+- Email template plugin settings now support aliases and .env variables.
+
+### Fixed
+- Fix an issue for case-sensitive email checks.
+
 ## 3.0.3 - 2024-11-21
 
 ### Fixed